More comments and some cleanups based on comments
- Rebase on latest arrow
- Remove VectorRoot and move to VectorSchemaRoot.
- Add AutoCloseables.
- Clean close invocation in VectorSchemaRoot
- Add more comments to proto
