+++ java/vector/src/main/java/org/apache/arrow/util/AutoCloseables.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.arrow.flight;
+package org.apache.arrow.util;
 
 import java.util.ArrayList;
 import java.util.Arrays;
