+++ java/vector/src/main/java/org/apache/arrow/vector/VectorSchemaRoot.java
@@ -22,10 +22,14 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import org.apache.arrow.memory.BufferAllocator;
+import org.apache.arrow.util.AutoCloseables;
 import org.apache.arrow.vector.types.pojo.Field;
 import org.apache.arrow.vector.types.pojo.Schema;
+
+import com.google.common.collect.ImmutableList;
 
 /**
  * Holder for a set of vectors to be loaded/unloaded
@@ -36,6 +40,19 @@ public class VectorSchemaRoot implements AutoCloseable {
   private int rowCount;
   private final List<FieldVector> fieldVectors;
   private final Map<String, FieldVector> fieldVectorsMap = new HashMap<>();
+
+
+  public VectorSchemaRoot(Iterable<FieldVector> vectors) {
+    this(
+        ImmutableList.copyOf(vectors).stream().map(t -> t.getField()).collect(Collectors.toList()),
+        ImmutableList.copyOf(vectors),
+        0
+        );
+  }
+
+  public VectorSchemaRoot(FieldVector...vectors) {
+    this(ImmutableList.copyOf(vectors));
+  }
 
   public VectorSchemaRoot(FieldVector parent) {
     this(parent.getField().getChildren(), parent.getChildrenFromFields(), parent.getValueCount());
@@ -73,8 +90,27 @@ public class VectorSchemaRoot implements AutoCloseable {
     return new VectorSchemaRoot(schema, fieldVectors, 0);
   }
 
+  /**
+   * Do an adaptive allocation of each vector for memory purposes. Sizes will be based on previously
+   * defined initial allocation for each vector (and subsequent size learnings).
+   */
+  public void allocateNew() {
+    for(FieldVector v : getFieldVectors()) {
+      v.allocateNew();
+    }
+  }
+
+  /**
+   * Release all the memory for each vector held in this root.
+   */
+  public void clear() {
+    for(FieldVector v : getFieldVectors()) {
+      v.clear();
+    }
+  }
+
   public List<FieldVector> getFieldVectors() {
-    return fieldVectors;
+    return ImmutableList.copyOf(fieldVectors);
   }
 
   public FieldVector getVector(String name) {
@@ -89,32 +125,27 @@ public class VectorSchemaRoot implements AutoCloseable {
     return rowCount;
   }
 
+  /**
+   * Set the row count of all the vectors in this container. Also sets the value count for each root level contained FieldVector.
+   * @param rowCount Number of records.
+   */
   public void setRowCount(int rowCount) {
     this.rowCount = rowCount;
+    for(FieldVector v : getFieldVectors()) {
+      v.setValueCount(rowCount);
+    }
   }
 
   @Override
   public void close() {
-    RuntimeException ex = null;
-    for (FieldVector fieldVector : fieldVectors) {
-      try {
-        fieldVector.close();
-      } catch (RuntimeException e) {
-        ex = chain(ex, e);
-      }
-    }
-    if (ex != null) {
+    try {
+      AutoCloseables.close(fieldVectors);
+    }catch(RuntimeException ex) {
       throw ex;
+    } catch (Exception ex) {
+      // should never happen since FieldVector.close() doesn't throw IOException
+      throw new RuntimeException(ex);
     }
-  }
-
-  private RuntimeException chain(RuntimeException root, RuntimeException e) {
-    if (root == null) {
-      root = e;
-    } else {
-      root.addSuppressed(e);
-    }
-    return root;
   }
 
   private void printRow(StringBuilder sb, List<Object> row) {
